字母配对修改

+++ capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/pojo/LetterPair.java
@@ -50,7 +50,7 @@ public class LetterPair extends Model<LetterPair> {
     /**
      * 记忆强度
      */
-    private Integer memoryStrength;
+    private Double memoryStrength;
 
     private Integer unitId;
